preview question logic
