--- packages/admin/src/pages/build-application/[applicationId]/[sectionId]/[questionId]/preview.module.scss
@@ -1,5 +1,0 @@
-@import '~govuk-frontend/govuk/helpers/colour';
-
-.gap-light-grey-background {
-    background-color: govuk-colour("light-grey");
-}
